Fix HTTP status checking, type safety, and error handling
- Add error_for_status() to all HTTP calls to catch non-2xx responses
- Fix reboot() to also tolerate body-read errors (is_body()) after the
  AP drops the connection
- Change SaveTagConfig.content_mode and UploadImageOptions.content_mode
  from Option<u8> to Option<ContentMode> for type-safe round-tripping
- Change ApListItem.count from u8 to u32 to handle >255 tags
- Replace unreachable mime_str error mapping with expect()
- Make check_response_body case-insensitive

+++ src/types/ap_item.rs
@@ -8,7 +8,7 @@ pub struct ApListItem {
     /// AP display name.
     pub alias: String,
     /// Number of tags managed by this AP.
-    pub count: u8,
+    pub count: u32,
     /// Radio channel (string-encoded).
     pub channel: String,
     /// Firmware version as 4-character hex string.
